Cobrar un ticket por separado

+++ app/src/main/java/com/bakery/dam/androidtpv/model/Oferta.java
@@ -1,6 +1,6 @@
 package com.bakery.dam.androidtpv.model;
 
-import java.util.List;
+import java.math.BigDecimal;
 
 /**
  * Created by DAM on 24/2/17.
@@ -22,7 +22,7 @@ public class Oferta {
 
     private String nombre;
 
-    private double precio;
+    private BigDecimal precio;
 
     public String getDescripcion() {
         return descripcion;
@@ -80,11 +80,11 @@ public class Oferta {
         this.nombre = nombre;
     }
 
-    public double getPrecio() {
+    public BigDecimal getPrecio() {
         return precio;
     }
 
-    public void setPrecio(double precio) {
+    public void setPrecio(BigDecimal precio) {
         this.precio = precio;
     }
 }
